working on the LLVM assembly production compiler new features fixes

+++ compiler/structures.h
@@ -91,7 +91,7 @@ struct expr {
     struct symbol *symbol;
     struct type *type;
 
-    unsigned short reg;
+    int reg;
 };
 
 struct type {
